requirements.txt
arrow==0.10.0
cairocffi==0.8.0
CairoSVG==1.0.19
cffi==1.9.1
click==6.7
contextlib2==0.5.4
cssselect==1.0.1
dill==0.2.6
-e git+https://github.com/b1naryth1ef/disco.git@530a79f1139a9e535df87f9e8823e60a57584c3d#egg=disco-py
# disco-py==0.0.11-rc.8
Distance==0.1.3
emoji==0.3.9
enum34==1.1.6
erlpack==0.3.2
Flask==0.12
future==0.16.0
futures==3.0.5
greenlet==0.4.11
holster==1.0.11
httplib2==0.10.3
humanize==0.5.1
inflection==0.3.1
influxdb==4.0.0
itsdangerous==0.24
Jinja2==2.9.5
lxml==3.7.2
markovify==0.5.4
MarkupSafe==0.23
oauth2client==3.0.0
oauthlib==2.0.1
olefile==0.44
packaging==16.8
Pillow==4.0.0
ply==3.8
protobuf==3.2.0
psycogreen==1.0
psycopg2==2.6.2
pyasn1==0.2.3
pyasn1-modules==0.0.8
pycparser==2.17
pygal==2.3.1
pyparsing==2.2.0
pyquery==1.2.17
python-dateutil==2.6.0
pytz==2016.10
PyYAML==3.12
raven==5.32.0
redis==2.10.5
regex==2017.2.8
requests-oauthlib==0.8.0
rsa==3.4.2
ruamel.ordereddict==0.4.9
ruamel.yaml==0.13.14
tinycss==0.4
typing==3.5.3.0
tzlocal==1.3
Unidecode==0.4.20
websocket-client==0.37.0
Werkzeug==0.11.15
wheel==0.24.0
xxhash==1.0.1
fuzzywuzzy==0.15.0
pytest==3.0.7
python-Levenshtein==0.12.0
-e git+https://github.com/b1naryth1ef/peewee.git@1b7d9f34d5f93464aa647cea4faab68cc45c1496#egg=peewee
gevent_inotifyx==0.1.1
datadog==0.16.0
